git-epic-creator/services/code_graph_ingestion_service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "code-graph-ingestion-service"
version = "0.1.0"
description = "FastAPI service that ingests repos and persists deterministic code graphs to Neo4j."
authors = [
    { name = "Vladyslav Avilov", email = "vladyslav.avilov@email.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.121.0",
    "uvicorn==0.38.0",
    "pydantic==2.12.3",
    "pydantic-settings==2.11.0",
    "structlog==25.5.0",
    "python-multipart==0.0.20",
    "httpx==0.28.1",
    "dulwich==0.24.1",
    "pathspec==0.12.1",
    # Keep compatible with `tree-sitter-cobol` (tree-sitter==0.25.1).
    "tree-sitter-language-pack==0.7.4",
    "safezipfile==0.1.3",
    "opentelemetry-api==1.39.1",
    "opentelemetry-sdk==1.39.1",
    "shared @ file:../shared",
    # COBOL Tree-sitter binding lives in-repo (services/tree_sitter_cobol_binding).
    # IMPORTANT: This version is not published to PyPI; install from local path.
    "tree-sitter-cobol @ file:../tree_sitter_cobol_binding",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]
